+++ test/sample-splitting/a.js
@@ -1,0 +1,8 @@
+function a() {
+    console.log('a');
+    import('./c').then(function(c) {
+        c();
+    })
+}
+
+a();