+++ test/sample-splitting/b.js
@@ -1,0 +1,8 @@
+function b() {
+    console.log('b');
+    import('./c').then(function(c) {
+        c();
+    })
+}
+
+b();